Create, update, get by id and delete books

+++ src/main/java/julia/books/domain/books/BooksRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BooksRepository extends PagingAndSortingRepository<BookEntity, Integer> {
+public interface BooksRepository extends PagingAndSortingRepository<BookEntity, Long> {
 }
